cleaning up relationship display with delete-relationship can checks, ensuring links are to the correct relationship and have the ability to delete

+++ resources/views/dioceses/show.blade.php
@@ -158,19 +158,27 @@
                 <ul>
                     @foreach($diocese->a_relationships as $a_relationship)
                         <li>
+                          @can('delete-relationship')
                             {!! Form::open(['method' => 'DELETE', 'route' => ['relationship.destroy', $a_relationship->id],'onsubmit'=>'return ConfirmDelete()']) !!}
                                 {!!$diocese->contact_link!!} {{ $a_relationship->relationship_type->label_a_b }} {!! $a_relationship->contact_b->contact_link !!}
                                 <button type="submit" class="btn btn-outline-dark btn-sm"><i class="fas fa-trash"></i></button>
                             {!! Form::close() !!}
+                          @else
+                            {!!$diocese->contact_link!!} {{ $a_relationship->relationship_type->label_a_b }} {!! $a_relationship->contact_b->contact_link !!}
+                          @endCan
                         </li>
                     @endforeach
 
                     @foreach($diocese->b_relationships as $b_relationship)
                         <li>
+                          @can('delete-relationship')
                             {!! Form::open(['method' => 'DELETE', 'route' => ['relationship.destroy', $b_relationship->id],'onsubmit'=>'return ConfirmDelete()']) !!}
-                                {!!$diocese->contact_link!!} is {{ $b_relationship->relationship_type->label_a_b }} {!! $b_relationship->contact_b->contact_link !!}
+                                {!!$diocese->contact_link!!} is {{ $b_relationship->relationship_type->label_b_a }} {!! $b_relationship->contact_a->contact_link !!}
                                 <button type="submit" class="btn btn-outline-dark btn-sm"><i class="fas fa-trash"></i></button>
                             {!! Form::close() !!}
+                          @else
+                            {!!$diocese->contact_link!!} is {{ $b_relationship->relationship_type->label_b_a }} {!! $b_relationship->contact_a->contact_link !!}
+                          @endCan
                         </li>
                     @endforeach
                 </ul>
